refactor(URM): follow #1279
